+++ src/cl.debug/CHANGELOG.rst
@@ -2,6 +2,15 @@
 Cl.Debug
 ========
 
+1.2.0
+-----
+- added zoom functionality
+- added default options for grid
+- changed bookmarklet to accept title
+- changed helpers to be centralized
+- removed not required designmode and resizable
+- removed firebug lite from list
+
 1.1.3
 -----
 - created class.js element out of cl.debug